Add validation and exception handling.

+++ src/main/java/com/example/rewardsrestful/controller/CustomerController.java
@@ -1,7 +1,12 @@
 package com.example.rewardsrestful.controller;
 
+import com.example.rewardsrestful.exception.CustomerNotFoundException;
+import com.example.rewardsrestful.impl.RewardServiceImpl;
 import com.example.rewardsrestful.model.Customer;
+import com.example.rewardsrestful.model.ResponseMessage;
 import com.example.rewardsrestful.service.CustomerService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -12,12 +17,13 @@ import org.springframework.web.bind.annotation.*;
 public class CustomerController {
 
     private CustomerService customerService;
+    private static Logger logger = LoggerFactory.getLogger(RewardServiceImpl.class);
 
     public CustomerController(CustomerService customerService) {
         this.customerService = customerService;
     }
 
-    @GetMapping("/customers")
+    @GetMapping("/customer")
     public ResponseEntity<?> getAllCustomers() {
 
         return new ResponseEntity<>(customerService.findAllCustomers(), HttpStatus.OK);
@@ -26,34 +32,51 @@ public class CustomerController {
     @GetMapping("/customer/{id}")
     public ResponseEntity<?> getCustomerById(@PathVariable("id") Long id) {
 
-        return new ResponseEntity<>(customerService.findCustomerById(id), HttpStatus.OK);
+        Customer customer = customerService.findCustomerById(id);
+        if (customer == null) {
+            throw new CustomerNotFoundException("Customer not found.");
+        }
+        return new ResponseEntity<>(customer, HttpStatus.OK);
     }
 
     @PostMapping("/customer")
     public ResponseEntity<?> createCustomer(@Validated @RequestBody Customer customer) {
 
         Customer createdCustomer = customerService.createCustomer(customer);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+        return new ResponseEntity<>(new ResponseMessage("Customer created.", createdCustomer), HttpStatus.CREATED);
     }
 
     @PutMapping("/customer/{id}")
     public ResponseEntity<?> updateCustomer(@PathVariable("id") Long id, @Validated @RequestBody Customer customer) {
 
         Customer updatedCustomer = customerService.findCustomerById(id);
+        if (updatedCustomer == null) {
+            throw new CustomerNotFoundException("Customer not found.");
+        }
         updatedCustomer.setEmail(customer.getEmail());
         updatedCustomer.setId(customer.getId());
         updatedCustomer.setName(customer.getName());
 
         customerService.updateCustomer(updatedCustomer);
-        return new ResponseEntity<>(HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(new ResponseMessage("Customer updated.", updatedCustomer), HttpStatus.OK);
     }
 
     @DeleteMapping("/customer/{id}")
     public ResponseEntity<?> deleteCustomer(@PathVariable("id") Long id) {
 
+        Customer deletedCustomer = customerService.findCustomerById(id);
+        if (deletedCustomer == null) {
+            throw new CustomerNotFoundException("Customer not found.");
+        }
+
         customerService.deleteCustomerById(id);
-        return new ResponseEntity<>(HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(new ResponseMessage("Customer deleted."), HttpStatus.OK);
     }
 
+    @ExceptionHandler(CustomerNotFoundException.class)
+    public ResponseEntity<?> exceptionHandlerCustomerNotFound(Exception ex) {
 
+        logger.error(ex.getMessage());
+        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+    }
 }
